Performing cleanup, added new class, still wip
